colive-distributed-middleware/requirements.txt
python-socketio[client]
flask
